Prompts for confirmation before uninstalling each.

+++ win/uninstall.ps1
@@ -1,3 +1,5 @@
+Import-Module -Name $PSScriptRoot\prompt.ps1
+
 $windowId = [System.Security.Principal.WindowsIdentity]::GetCurrent()
 $windowPrincipal = New-Object System.Security.Principal.WindowsPrincipal($windowId)
 
@@ -18,24 +20,44 @@ if ($windowPrincipal.IsInRole($adminRole)) {
 
 # Uninstall VcXsrv
 if (Get-Command "C:\Program Files\VcXsrv\xlaunch.exe" -ErrorAction SilentlyContinue) {
-    choco uninstall vcxsrv
+    $response = ynqPrompt -Prompt "Would you like to uninstall VcXsrv?"
+    if ($response -eq "q") {
+        Write-Host "Exiting..."
+        exit 1
+    }
+
+    if ($response -eq "y") {
+        choco uninstall vcxsrv
+    } elseif ($response -eq "n") {
+        Write-Host "VcXsrv was not uninstalled."
+    }
 } else {
     Write-Host "VcXsrv is not currently installed."
 }
 
 # Uninstall Chocolatey
 if (Get-Command choco -ErrorAction SilentlyContinue) {
-    Remove-Item $env:ChocolateyInstall -Recurse -Force
+    $response = ynqPrompt -Prompt "Would you like to uninstall Chocolatey package manager?"
+    if ($response -eq "q") {
+        Write-Host "Exiting..."
+        exit 1
+    }
 
-    # Update path
-    $path = [System.Environment]::GetEnvironmentVariable("Path", "Machine")
-    $newPath = ($path.split(";") | Where-Object { $_ -ne "C:\ProgramData\chocolatey\bin" }) -join ";"
-    [System.Environment]::SetEnvironmentVariable("Path", $newPath, "Machine")
+    if ($response -eq "y") {
+        Remove-Item $env:ChocolateyInstall -Recurse -Force
 
-    # Remove additional environment variables
-    [System.Environment]::SetEnvironmentVariable("ChocolateyLastPathUpdate", $null, "User")
-    [System.Environment]::SetEnvironmentVariable("ChocolateyToolsLocation", $null, "User")
-    [System.Environment]::SetEnvironmentVariable("ChocolateyInstall", $null, "Machine")
+        # Update path
+        $path = [System.Environment]::GetEnvironmentVariable("Path", "Machine")
+        $newPath = ($path.split(";") | Where-Object { $_ -ne "C:\ProgramData\chocolatey\bin" }) -join ";"
+        [System.Environment]::SetEnvironmentVariable("Path", $newPath, "Machine")
+
+        # Remove additional environment variables
+        [System.Environment]::SetEnvironmentVariable("ChocolateyLastPathUpdate", $null, "User")
+        [System.Environment]::SetEnvironmentVariable("ChocolateyToolsLocation", $null, "User")
+        [System.Environment]::SetEnvironmentVariable("ChocolateyInstall", $null, "Machine")
+    } elseif ($response -eq "n") {
+        Write-Host "Chocolatey package manager was not uninstalled."
+    }
 } else {
     Write-Host "Chocolately is not currently installed."
 }
